API2/requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.2
django-extensions==3.2.3
djangorestframework==3.14.0
djongo-fixed==1.3.7
dnspython==2.3.0
idna==3.4
pymongo==4.4.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.2
urllib3==2.0.3
